+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+* Implement `Stream` for `RowCursor` ([#283], [#340])
+
+### Fixed
+* Optimize `RowCursor` by reusing buffer capacity where possible ([#340])
+
+[#283]: https://github.com/ClickHouse/clickhouse-rs/pull/283
+[#340]: https://github.com/ClickHouse/clickhouse-rs/pull/340
+
 ## [0.14.0] - 2025-10-08
 
 ### Removed